pyproject.toml
[project]
name = "azure-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "~=3.13"
dependencies = [
    "azure-identity>=1.25.1",
    "azure-search-documents>=11.6.0",
    "fastexcel>=0.18.0",
    "ipython>=9.8.0",
    "langchain>=1.0.3",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.0.2",
    "langgraph>=1.0.2",
    "langgraph-cli[inmem]>=0.4.7",
    "notebook>=7.5.0",
    "polars>=1.35.2",
    "python-dotenv>=1.2.1",
]
